+++ client/helpers/router.js
@@ -117,7 +117,7 @@ var filters = {
 
   isLoggedIn: function() {
     if (!(Meteor.loggingIn() || Meteor.user())) {
-      throwError('Please Sign In First.')
+      throwError(i18n.t('Please Sign In First.'))
       this.render('signin');
       this.stop(); 
     }
@@ -132,7 +132,7 @@ var filters = {
 
   isAdmin: function() {
     if(!Meteor.loggingIn() && Session.get('settingsLoaded') && !isAdmin()){
-      throwError("Sorry, you  have to be an admin to view this page.")
+      throwError(i18n.t("Sorry, you  have to be an admin to view this page."))
       this.render('no_rights');
       this.stop(); 
     }
@@ -148,7 +148,7 @@ var filters = {
 
   canPost: function () {
     if(!Meteor.loggingIn() && Session.get('settingsLoaded') && !canPost()){
-      throwError("Sorry, you don't have permissions to add new items.")
+      throwError(i18n.t("Sorry, you don't have permissions to add new items."))
       this.render('no_rights');
       this.stop();      
     }
@@ -157,7 +157,7 @@ var filters = {
   canEditPost: function() {
     var post = Posts.findOne(this.params._id);
     if(!Meteor.loggingIn() && Session.get('settingsLoaded') && !currentUserCanEdit(post)){
-      throwError("Sorry, you cannot edit this post.")
+      throwError(i18n.t("Sorry, you cannot edit this post."))
       this.render('no_rights');
       this.stop();
     }
@@ -166,7 +166,7 @@ var filters = {
   canEditComment: function() {
     var comment = Comments.findOne(this.params._id);
     if(!Meteor.loggingIn() && Session.get('settingsLoaded') && !currentUserCanEdit(comment)){
-      throwError("Sorry, you cannot edit this comment.")
+      throwError(i18n.t("Sorry, you cannot edit this comment."))
       this.render('no_rights');
       this.stop();
     }
@@ -249,30 +249,23 @@ PostsListController = RouteController.extend({
   waitOn: function () {
     // take the first segment of the path to get the view, unless it's '/' in which case the view default to 'top'
     // note: most of the time this.params.slug will be empty
-    var view = this.path == '/' ? 'top' : this.path.split('/')[1],
-        limit = this.params.limit || getSetting('postsPerPage', 10),
-        parameters = getParameters(view, limit, this.params.slug, Session.get("searchQuery"));
-        
+    this._terms = {
+      view: this.path == '/' ? 'top' : this.path.split('/')[1],
+      limit: this.params.limit || getSetting('postsPerPage', 10),
+      category: this.params.slug,
+      query: Session.get("searchQuery")
+    }
     return [
-      Meteor.subscribe('postsList', parameters.find, parameters.options),
-      Meteor.subscribe('postsListUsers', parameters.find, parameters.options)
+      Meteor.subscribe('postsList', this._terms),
+      Meteor.subscribe('postsListUsers', this._terms)
     ]
   },
   data: function () {
-    var view = this.path == '/' ? 'top' : this.path.split('/')[1],
-        limit = this.params.limit || getSetting('postsPerPage', 10),
-        parameters = getParameters(view, limit, this.params.slug, Session.get("searchQuery")),
+    var parameters = getParameters(this._terms),
         posts = Posts.find(parameters.find, parameters.options);
         postsCount = posts.count();
   
-    Session.set('postsLimit', limit);
-
-    // get posts and decorate them with rank property
-    // note: not actually used; 
-    // posts = posts.map(function (post, index) {
-    //   post.rank = index;
-    //   return post;
-    // });
+    Session.set('postsLimit', this._terms.limit);
 
     return {
       postsList: posts,
@@ -291,16 +284,25 @@ PostsDigestController = RouteController.extend({
   template: 'posts_digest',
   waitOn: function() {
     // if day is set, use that. If not default to today
-    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today');
-    var parameters = getDigestParameters(currentDate);
+    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today'),
+        terms = {
+          view: 'digest',
+          after: moment(currentDate).startOf('day').valueOf(),
+          before: moment(currentDate).endOf('day').valueOf()
+        }
     return [
-      Meteor.subscribe('postsList', parameters.find, parameters.options),
-      Meteor.subscribe('postsListUsers', parameters.find, parameters.options)
+      Meteor.subscribe('postsList', terms),
+      Meteor.subscribe('postsListUsers', terms)
     ]
   },
   data: function() {
-    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today');
-    var parameters = getDigestParameters(currentDate);
+    var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today'),
+        terms = {
+          view: 'digest',
+          after: moment(currentDate).startOf('day').valueOf(),
+          before: moment(currentDate).endOf('day').valueOf()
+        },
+        parameters = getParameters(terms);
     Session.set('currentDate', currentDate);
     return {
       posts: Posts.find(parameters.find, parameters.options)
@@ -358,8 +360,13 @@ UserPageController = RouteController.extend({
   data: function() {
     var findById = Meteor.users.findOne(this.params._idOrSlug);
     var findBySlug = Meteor.users.findOne({slug: this.params._idOrSlug});
-    return {
-      user: (typeof findById == "undefined") ? findBySlug : findById
+    if(typeof findById !== "undefined"){
+      // redirect to slug-based URL
+      Router.go(getProfileUrl(findById), {replaceState: true});
+    }else{
+      return {
+        user: (typeof findById == "undefined") ? findBySlug : findById
+      }
     }
   }
 });
@@ -537,9 +544,8 @@ Router.map(function() {
     path: '/all-users/:limit?',
     template: 'users',
     waitOn: function() {
-      var limit = parseInt(this.params.limit) || 20,
-          parameters = getUsersParameters(this.params.filterBy, this.params.sortBy, limit);
-      return Meteor.subscribe('allUsers', parameters.find, parameters.options);
+      var limit = parseInt(this.params.limit) || 20;
+      return Meteor.subscribe('allUsers', this.params.filterBy, this.params.sortBy, limit);
     },
     data: function() {
       var limit = parseInt(this.params.limit) || 20,