+++ History.md
@@ -1,6 +1,15 @@
+## v0.7.4 “InterScope”
+
+* Added basic internationalization (thanks Toam!). 
+
+## v0.7.3
+
+* Refactored notifications.
+* Added notifications for new users creation.
+
 ## v0.7.2
 
-* Added basic search (thanks Ry!)
+* Added basic search (thanks Ry!).
 
 ## v0.7.1